👌 IMPROVE: City refactor
