+++ utils/green.js
@@ -1,0 +1,2 @@
+const chalk = require('chalk');
+module.exports = string => chalk.hex(`#81EF96`)(string);
